// src/types.ts
export type Screen =
    | 'splash'
    | 'welcome'
    | 'login'
    | 'signup'
    | 'kycPrompt'
    | 'kycForm'
    | 'kycPending'
    | 'wallet'
    | 'cardDesign'
    | 'transactions'
    | 'deposit'
    | 'withdraw'
    | 'mt5Account'
    | 'trades'
    | 'home'
    | 'transfer'
    | 'plans'
    | 'settings'
    | 'security'
    | 'accountSettings'
    | 'personalData'
    | 'forgotPassword'
    | 'verifyPassword'
    | 'resetPassword';              